Update models in data and domain modules

+++ data/src/main/java/com/richard/data/racing/model/RaceFormModel.kt
@@ -1,7 +1,7 @@
 package com.richard.data.racing.model
 
 class RaceFormModel(
-    val distance: Long,
+    val distance: Int,
     val distanceType: DistanceTypeModel,
     val trackCondition: TrackConditionModel,
     val venueId: VenueModel,
